+++ src/index.ts
@@ -1,5 +1,5 @@
 import {readFileSync} from "fs"
-import {ActivityType, ApplicationCommandType, InteractionType, PresenceUpdateStatus} from "discord.js";
+import {ActivityType, ApplicationCommandType, ContextMenuCommandBuilder, InteractionType, PresenceUpdateStatus} from "discord.js";
 import { ChatGPTBotClient } from "./classes/client";
 import { handleCommands } from "./handlers/commandHandler";
 import { handleComponents } from "./handlers/componentHandler";
@@ -53,7 +53,15 @@ client.on("ready", async () => {
     //client.modals.loadClasses().catch(console.error)
     client.user?.setPresence({activities: [{type: ActivityType.Listening, name: "to ChatGPT screaming at your requests"}], status: PresenceUpdateStatus.DoNotDisturb })
     console.log(`Ready`)
-    await client.application?.commands.set([...client.commands.createPostBody(), ...client.contexts.createPostBody()]).catch(console.error)
+    const configurable_msg_cmds = client.config.message_context_actions?.map((a, i) => 
+        new ContextMenuCommandBuilder()
+            .setType(ApplicationCommandType.Message)
+            .setName(a.name ?? `Unknown Action ${i}`)
+            .setDMPermission(false)
+            .toJSON()
+
+    ) || []
+    await client.application?.commands.set([...client.commands.createPostBody(), ...client.contexts.createPostBody(), ...configurable_msg_cmds]).catch(console.error)
 
     if(client.config.selectable_system_instructions?.length && client.config.selectable_system_instructions.some(i => !i.name || !i.system_instruction)) throw new Error("Every selectable system instruction needs a name and a system_instruction")
     if(client.config.selectable_system_instructions?.length && client.config.selectable_system_instructions.some(i => i.name === "default")) throw new Error("You can't name your system instruction 'default'")