+++ src/types.ts
@@ -164,14 +164,15 @@ export interface Config {
         regenerate_button?: boolean,
         delete_button?: boolean,
         view_system_instruction?: boolean,
-        englishify?: boolean,
-        context_action?: boolean,
         user_stats?: boolean,
         user_leaderboard?: boolean
     },
     leaderboard_amount_users?: number,
     englishify_system_instruction?: string,
-    context_action_instruction?: string,
+    message_context_actions?: {
+        name?: string,
+        system_instruction?: string
+    }[],
     costs?: Record<string, {
         prompt?: number,
         completion?: number